eval.l: instantiate-subrs is now handled by a macro
no need to list subrs twice anymore, just a single define-subr. this
allows to chop eval.l into mutliple files in the future to compile in
different sets of subrs (think of compiling on a platform without libc,
e.g. directly to a ZX80 and using its bios io routines)