major refactor of types, type-of now returns maru objects
define-structure -> define-record

introduces:
 - <type>, <record>
 - define-type to define types whose instances have no slots
 - subtype?
 - direct and effective slots, inspired by CLOS MOP

as a sideffect of this change the types can now finally be freely
reordered between the host and the slave, and their slot layout
may also be changed completely.

explicit errors are raised when the slave accesses an instance of
a host type that is not expected (i.e. stuff returned by the
host's reader will be dispatched on while bootstrapping).