makefile: reduce the 3-way bootstrap to a 2-way one
looks like whenever a substantial change would justify the old
eval1 step, then it's a simpler way to solve that by developing
the two stages in parallel until the bootstrap works again.

this is more general than the old setup that only worked in special
situations, e.g. when eval.l wasn't relying much on the new feature,
only the compiler, and/or eval.l had to be filled with dispatching
on the 'evolving? variable.