rework the slave-target isolation; eliminate *syntax-module* from eval.
there are three modules: host, slave, target.

it works, but the details still need to be worked out. it feels as
if the infrastructure was more complicated than it should be, but the
result is nice: eval.l can redefine e.g. eval with a function meant to
be level-shifted. the target even has now its own true and false
constants, mapped to 1 and 0.

the target module now is an empty module, and its entire env gets
level-shifted to the target. <form>s get copied from the slave to
the target, and forms defined in the target will be encoded in the
slave, which enables them to use all the bindings available in the
slave to construct their expansion.