the linux platform can bootstrap itself for the first time
it works, but it needs cleanup, too...

a lot of gc/let* is missing, especially with all the string-from-cstring,
but next should be to compile literal strings into proper, but read-only
heap objects. that will resolve all of those.

llvm on safety 1 sigsegv's, but runs through the bootstrap on safety 3.