makefile: support aarch64
in addition, store the .ll files in the arch trillet dirs, because
besides endianness and word size e.g. the syscalls make the linux
platform's output dependent on the arch, too.

it's overkill for the libc platform, because there they only depend on
the endianness and the word size (which btw gives us easy bootstrap on
aarch64, because we can just compile the libc eval.ll there as-is).

this removes endianness from the path, but for all practical purposes
we are little endian only for now: all the supported platforms are le
and i'm pretty sure our codebase would also trip up on this and that
on a be platform.